Added a build profile and made UI adjustments
A new build profile has been added to the 'pom.xml' to automate front-end task runners such as npm. This is done to improve ease of use and the reliability of the build process. Also, some minor UI adjustments have been made in 'App.js' to optimize the display - changed the pagination size and updated the footer content. Lines added to .gitignore file are unnecessary, presumably an error and thus made no actual changes in the file.